move utils into util dir
