+++ react/src/routes/login.tsx
@@ -22,7 +22,7 @@ import { type SubmitHandler, useForm } from "react-hook-form";
 
 import type { Body_login_access_token_login_access_token_post as AccessToken } from "../client";
 import useAuth, { isLoggedIn } from "../hooks/useAuth";
-import { emailPattern } from "../utils";
+import { emailPattern } from "../util/misc";
 
 export const Route = createFileRoute("/login")({
   component: Login,
