+++ react/src/util/misc.ts
